Fixing bug in CakePHP 2.1
2.1 the find first is not return false it is returning one empty array

+++ Model/Behavior/UploadBehavior.php
@@ -567,7 +567,7 @@ class UploadBehavior extends ModelBehavior
             ),
         );
 
-        if ($attachment !== false) {
+        if (!empty($attachment) && $attachment !== false) {
             $this->deleteAllFiles($model, $attachment);
             $data[$className]['id'] = $attachment[$className]['id'];
         } else {
